[AtmanOS PoC]: lots of status/debugging tweaks
The messages emitted by the module were not very useful or
informative; make them summarize what happened a bit.

If you're wondering about the Noop() functions in this module, this
was so that I could be sure that the packages were being built into
the test program and therefore the relevant init() functions being
called.

ARP messages are now processed correctly and logged as:

    TRACE: 03:26:04.147254 arp.go:65: who-is 10.0.2.42?  I am!
    TRACE: 03:34:50.439185 arp.go:89: who-is 10.0.2.41?  not me, I'm 10.0.2.42

TCP SYNs start handler goroutines:

    TRACE: 03:25:32.918037 ip_read_manager.go:71: IP Proto 06, 10.0.2.15 -> 10.0.2.42
    mpreinit(0xc820019400)
    newosproc stk=0xc820141fc0 m=0xc820019400 g=0xc82013e600 id=4 ostk=0xc820025e70
    minit()
    ERROR: 03:25:33.418255 tcp_server.go:105: ARP request timed out